Send response header asap.
Previously header is saved along with body in the first cache slot,
when the response is greater than bufsize, a memory zone of bufzie
will be allocated for the first slot and the most of it will be filled.
If `rspadd` or `http-response set-header` is set in the conf file,
and if the header is long, there will be no room to accommodate the
headers and leads to hang.

This fix separate header and body, and send header along.